Add `yii\base\MutatorsTrait`.
