+++ framework/base/MutatorsInterface.php
@@ -1,0 +1,73 @@
+<?php
+/**
+ * @link http://www.yiiframework.com/
+ * @copyright Copyright (c) 2008 Yii Software LLC
+ * @license http://www.yiiframework.com/license/
+ */
+
+namespace yii\base;
+
+/**
+ *
+ * @author Robert Korulczyk <robert@korulczyk.pl>
+ * @since 2.0.11
+ */
+interface MutatorsInterface
+{
+
+    /**
+     * Returns a value indicating whether a property is defined.
+     * A property is defined if:
+     *
+     * - the class has a getter or setter method associated with the specified name
+     *   (in this case, property name is case-insensitive);
+     * - the class has a member variable with the specified name (when `$checkVars` is true);
+     *
+     * @param string $name the property name
+     * @param boolean $checkVars whether to treat member variables as properties
+     * @return boolean whether the property is defined
+     * @see canGetProperty()
+     * @see canSetProperty()
+     */
+    public function hasProperty($name, $checkVars = true);
+
+    /**
+     * Returns a value indicating whether a property can be read.
+     * A property is readable if:
+     *
+     * - the class has a getter method associated with the specified name
+     *   (in this case, property name is case-insensitive);
+     * - the class has a member variable with the specified name (when `$checkVars` is true);
+     *
+     * @param string $name the property name
+     * @param boolean $checkVars whether to treat member variables as properties
+     * @return boolean whether the property can be read
+     * @see canSetProperty()
+     */
+    public function canGetProperty($name, $checkVars = true);
+
+    /**
+     * Returns a value indicating whether a property can be set.
+     * A property is writable if:
+     *
+     * - the class has a setter method associated with the specified name
+     *   (in this case, property name is case-insensitive);
+     * - the class has a member variable with the specified name (when `$checkVars` is true);
+     *
+     * @param string $name the property name
+     * @param boolean $checkVars whether to treat member variables as properties
+     * @return boolean whether the property can be written
+     * @see canGetProperty()
+     */
+    public function canSetProperty($name, $checkVars = true);
+
+    /**
+     * Returns a value indicating whether a method is defined.
+     *
+     * The default implementation is a call to php function `method_exists()`.
+     * You may override this method when you implemented the php magic method `__call()`.
+     * @param string $name the method name
+     * @return boolean whether the method is defined
+     */
+    public function hasMethod($name);
+}
